Refactor namespaces from `Wangkanai.Caster.Identity` to `Wangkanai.Caster.Persistence` across the solution for enhanced module alignment.
