+++ src/Server/Components/Account/Pages/ExternalLogin.razor
@@ -6,7 +6,7 @@
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject SignInManager<CasterUser> SignInManager
 @inject UserManager<CasterUser>   UserManager
