+++ src/Server/Components/Account/Pages/Login.razor
@@ -3,7 +3,7 @@
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject SignInManager<CasterUser> SignInManager
 @inject ILogger<Login>                 Logger
