+++ src/Server/Components/Account/Pages/LoginWithRecoveryCode.razor
@@ -2,7 +2,7 @@
 
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject SignInManager<CasterUser> SignInManager
 @inject UserManager<CasterUser>   UserManager
