+++ src/Server/Components/Account/Pages/Manage/ChangePassword.razor
@@ -2,7 +2,7 @@
 
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>   UserManager
 @inject SignInManager<CasterUser> SignInManager
