+++ src/Server/Components/Account/Pages/Manage/EnableAuthenticator.razor
@@ -5,7 +5,7 @@
 @using System.Text
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser> UserManager
 @inject IdentityUserAccessor         UserAccessor
