+++ src/Server/Components/Account/Pages/Manage/ExternalLogins.razor
@@ -2,7 +2,7 @@
 
 @using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>   UserManager
 @inject SignInManager<CasterUser> SignInManager
