+++ src/Server/Components/Account/Pages/Manage/GenerateRecoveryCodes.razor
@@ -1,7 +1,7 @@
 ﻿@page "/Account/Manage/GenerateRecoveryCodes"
 
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>   UserManager
 @inject IdentityUserAccessor           UserAccessor
