+++ src/Server/Components/Account/Pages/Manage/ResetAuthenticator.razor
@@ -1,7 +1,7 @@
 ﻿@page "/Account/Manage/ResetAuthenticator"
 
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>   UserManager
 @inject SignInManager<CasterUser> SignInManager
