+++ src/Server/Components/Account/Pages/Manage/TwoFactorAuthentication.razor
@@ -2,7 +2,7 @@
 
 @using Microsoft.AspNetCore.Http.Features
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>   UserManager
 @inject SignInManager<CasterUser> SignInManager
