+++ src/Server/Components/Account/Pages/RegisterConfirmation.razor
@@ -3,7 +3,7 @@
 @using System.Text
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>  UserManager
 @inject IEmailSender<CasterUser> EmailSender
