+++ src/Server/Components/Account/Pages/ResendEmailConfirmation.razor
@@ -5,7 +5,7 @@
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject UserManager<CasterUser>  UserManager
 @inject IEmailSender<CasterUser> EmailSender
