+++ src/Server/Components/Account/Pages/ResetPassword.razor
@@ -4,7 +4,7 @@
 @using System.Text
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject IdentityRedirectManager      RedirectManager
 @inject UserManager<CasterUser> UserManager
