+++ src/Server/Components/Account/Shared/ExternalLoginPicker.razor
@@ -1,6 +1,6 @@
 ﻿@using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject SignInManager<CasterUser> SignInManager
 @inject IdentityRedirectManager        RedirectManager
