+++ src/Server/Components/Account/Shared/ManageNavMenu.razor
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using Wangkanai.Caster.Identity
+@using Wangkanai.Caster.Persistence
 
 @inject SignInManager<CasterUser> SignInManager
