+++ src/Server/Usings.cs
@@ -1,4 +1,3 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-global using Microsoft.EntityFrameworkCore;
 global using Wangkanai.Caster.Identity;
